Reorganize module_utils into common/ and rest/ subdirectories
Move files into logical subdirectories:
- common/pydantic_compat.py
- rest/response_handler_nd.py, rest_send.py, results.py, sender_nd.py
- rest/protocols/response_handler.py, response_validation.py, sender.py
- rest/response_strategies/nd_v1_strategy.py

Update all imports in moved files, nd_v2.py, and unit tests.
